requirements.txt
apache-airflow==2.6.1
apache-airflow-providers-http==3.4.0
pandas==2.0.3
s3fs==2023.1.0
python-dotenv